Acceptera bråkdels-kartpositioner i map-schemat (int -> float)
ScenePosition krävde x/y som int, så en drag till sub-pixel (eller en
befintlig map.json med bråkdels-positioner, t.ex. viksjo-kyrkogard) gav
Pydantic int_from_float och blockerade HELA POST /map -> stort 422-fel,
ingen kartsparning möjlig. Byt till float (positioner får vara sub-pixel;
prickarna renderas ändå i procent av naturlig storlek). Verifierat
end-to-end: viksjo-kyrkogards map.json (43 scener, bråktal) sparas nu 200.

+++ app/schemas.py
@@ -5,8 +5,11 @@ from pydantic import BaseModel, ConfigDict, Field
 
 
 class ScenePosition(BaseModel):
-    x: int
-    y: int
+    # Naturliga kartpixlar - får vara sub-pixel (drag ger bråktal, och prickarna
+    # renderas ändå i procent av naturlig storlek). int-krav avvisade annars både
+    # nya drag OCH befintliga map.json med bråkdels-positioner (int_from_float).
+    x: float
+    y: float
 
 
 class MapScene(BaseModel):
